Move 'netip' utils from 'ipam' to 'internal'.
